requirements.txt
numpy==1.23.5
pandas==1.5.2
python-dotenv==1.0.1
requests==2.31.0
scipy==1.9.1
yahoo_fin==0.8.9.1
yfinance==0.2.36
